cleanup after implementing autogeneration of IDs

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/InStoreOrderRepository.java
@@ -9,7 +9,7 @@ import ca.mcgill.ecse321.grocerystore.model.InStoreOrder;
 
 public interface InStoreOrderRepository extends CrudRepository<InStoreOrder, Integer> {
 
-	InStoreOrder findByOrderID(Integer orderID);
+	InStoreOrder findByOrderID(Long orderID);
 	
 	List<InStoreOrder> findByAccount(Account account);
 }
