+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/dao/WorkingHourRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import ca.mcgill.ecse321.grocerystore.model.WorkingHour;
 
 public interface WorkingHourRepository extends CrudRepository<WorkingHour, Integer> {
-	WorkingHour findByWorkingHourID(Integer workingHourID);
+	
+	WorkingHour findByWorkingHourID(Long workingHourID);
 }
